+++ src/test/java/roomescape/reservation/ReservationRepositoryTest.java
@@ -2,39 +2,20 @@ package roomescape.reservation;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import io.restassured.RestAssured;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.test.annotation.DirtiesContext;
+import roomescape.TestSupport;
 import roomescape.reservation.domain.Reservation;
 import roomescape.reservation.repository.ReservationRepository;
 import roomescape.time.domain.ReservationTime;
 import roomescape.time.repository.TimeRepository;
 
+public class ReservationRepositoryTest extends TestSupport {
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-public class ReservationRepositoryTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void beforeEach() {
-        RestAssured.port = port;
-    }
-
-    @Autowired
-    private JdbcTemplate jdbcTemplate;
     @Autowired
     private ReservationRepository reservationRepository;
     @Autowired
