+++ src/test/java/roomescape/time/ReservationTimePageControllerTest.java
@@ -1,25 +1,11 @@
 package roomescape.time;
 
 import io.restassured.RestAssured;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.annotation.DirtiesContext;
+import roomescape.TestSupport;
 
-@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
-@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationTimePageControllerTest {
-
-    @LocalServerPort
-    private int port;
-
-    @BeforeEach
-    void beforeEach() {
-        RestAssured.port = port;
-    }
+class ReservationTimePageControllerTest extends TestSupport {
 
     @Test
     @DisplayName("/startAt 을 요청하면 startAt.html 를 반환한다.")
@@ -29,5 +15,4 @@ class ReservationTimePageControllerTest {
                 .then().log().all()
                 .statusCode(200);
     }
-
 }
